requirements.txt
brevo-python>=4.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
